requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
av==9.0.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blessings==1.7
boto3==1.24.62
botocore==1.27.62
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
deepface==0.0.74
defusedxml==0.7.1
easydict==1.9
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
facenet-pytorch==2.5.2
fer==22.4.0
ffmpeg-python==0.2.0
filelock==3.6.0
fire==0.4.0
flake8==4.0.1
Flask==2.1.1
flatbuffers==1.12
fonttools==4.29.1
future==0.18.2
fvcore==0.1.5.post20220305
gast==0.4.0
gdown==4.4.0
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.20.4
humanfriendly==10.0
idna==3.3
imageio==2.16.1
img2vec-pytorch==1.0.1
importlib-metadata==4.11.1
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mtcnn==0.1.1
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.1
nvidia-ml-py3==7.352.0
nvtx==0.2.5
oauth2client==4.1.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.1
portalocker==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.4
pycodestyle==2.8.0
pycparser==2.21
pyee==9.0.3
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.2.1
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.10
rsa==4.8
s3transfer==0.6.0
sahi==0.10.4
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.1
sparkmeasure==0.14.0
stack-data==0.2.0
tabulate==0.8.9
tb-nightly==2.9.0a20220313
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
terminaltables==3.1.10
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
yacs==0.1.8
yapf==0.32.0
yolov5==6.1.8
youtube-dl==2021.12.17
zipp==3.7.0
